% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggIndex.R
\name{aggIndex}
\alias{aggIndex}
\title{Create a full index object by hedonic approach}
\usage{
aggIndex(trans_df, ...)
}
\arguments{
\item{trans_df}{data.frame of transactions}

\item{...}{Additional Arguments}
}
\value{
`hpi`` object.  S3 list with:
\describe{
\item{data}{`hpidata` object}
\item{model}{`hpimodel` object}
\item{index}{`hpiindex` object}
}
}
\description{
Wrapper to create index object via entire hedonic approach
}
\section{Further Details}{

Additional argument need to provide necessary argument for create `hpidata` objects if
the `trans_df` object is not of that class.
}

\examples{

 # Load data
 data(ex_sales)

 # Create index with raw transaction data
 agg_index <- aggIndex(trans_df = ex_sales,
                       estimator = 'median')

}
