% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggModel.R
\name{aggModel.mean}
\alias{aggModel.mean}
\title{Hedonic model approach with base estimator}
\usage{
\method{aggModel}{mean}(estimator, hed_df, price_field = "price", ...)
}
\arguments{
\item{estimator}{Type of model to estimates (base, robust, weighted)}

\item{hed_df}{Repeat sales dataset from hedCreateSales()}

\item{price_field}{Field on which to calcualate aggregat index)}

\item{...}{Additional arguments}
}
\description{
Use of base estimator in hedonic model approach
}
\section{Further Details}{

See `?aggModel` for more information
}

