% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blendIndexes.R
\name{blendIndexes}
\alias{blendIndexes}
\title{blendIndexes}
\usage{
Lorem Ipsum...
}
\arguments{
\item{index_list}{A list of identical length indexes}

\item{weights}{default=NULL; A vector of weights}

\item{...}{Additional Arguments}
}
\value{
a ts object with blended index
}
\description{
Blend together two or more indexes
}
\section{Further Details}{

Leaving weights to be NULL results in a 1/n weighting.
}

\examples{
a <- 1
}
