% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcForecastError.R
\name{buildForecastIDs}
\alias{buildForecastIDs}
\title{buildForecastIDs}
\usage{
buildForecastIDs(time_cut, hpi_df, ...)
}
\arguments{
\item{time_cut}{Period after which to cut off data}

\item{hpi_df}{Data to be converted to training or scoring}

\item{train}{Default=TRUE; Create training data?  FALSE = Scoring data}

\item{...}{Additional Arguments}
}
\value{
vector of row_ids indicating inclusion in the forecasting data as either the training
set (train = TRUE) or the scoring set (train = FALSE)
}
\description{
Create training or scoring data for the forecast error calculations
}
\section{Further Details}{

This function is rarely (if ever) used directly.  Most often called by `calcForecastError()`
}

\examples{
\dontrun{
buildForecastIDs(time_cut, hpi_df, ...)
}
}
