% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcForecastErrors.R
\name{calcForecastErrors}
\alias{calcForecastErrors}
\title{calcForecastErrors}
\usage{
Lorem Ipsum...
}
\arguments{
\item{pred_data}{Set of sales to be used for predicitive quality of index}

\item{...}{Additional Arguments}

\item{hpi_obj}{Object of class 'hpi'}

\item{train_range}{Number of periods to use as purely training before forecast starts}

\item{max_period}{Default=NULL; Maximum number of periods to forecast up to}
}
\value{
hpimodel object
}
\description{
Estimate out-of-sample index errors using a forecast method
}
\section{Further Details}{

If 'max_period' is left NULL, then it will forecast up to the end of the data
}

\examples{
a <- 1
}
