% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHPIError.R
\name{calcHPIError}
\alias{calcHPIError}
\title{calcHPIError}
\usage{
Lorem Ipsum...
}
\arguments{
\item{pred_data}{Set of sales against which to test predictions}

\item{index}{Index to be tested for accuracy}

\item{...}{Additional Arguments}
}
\value{
hpimodel object
}
\description{
Estimate the predictive error of an index. Generic method.
}
\section{Further Details}{

Note that ensure prediction types (holdout vs forecast) are done outside of this function
}

\examples{
calcHPIError(pred_data=rs_data,
             index=rs_index)
}
