% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIndexSeries.R
\name{calcIndexSeries}
\alias{calcIndexSeries}
\title{calcIndexSeries}
\usage{
Lorem Ipsum...
}
\arguments{
\item{hpi_obj}{Object of class 'hpi'}

\item{train_range}{Number of periods to use as purely training before forecast starts}

\item{max_period}{Default=NULL; Maximum number of periods to forecast up to}

\item{...}{Additional Arguments}

\item{pred_data}{Set of sales to be used for predicitive quality of index}
}
\value{
hpimodel object
}
\description{
Calculates a series of indexes, each one period longer than the previous
}
\section{Further Details}{

If 'max_period' is left NULL, then it will forecast up to the end of the data
}

\examples{
a <- 1
}
