% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcInSampleError.R
\name{calcInSampleError}
\alias{calcInSampleError}
\title{calcInSampleError}
\usage{
calcInSampleError(pred_df, index)
}
\arguments{
\item{pred_df}{Set of sales against which to test predictions}

\item{index}{Index (of class) to be tested for accuracy}

\item{...}{Additional Arguments}
}
\value{
object of class `indexerrors` inheriting from class `data.frame` containing the following fields:
\item{prop_id}
\item{pred_price}
\item{pred_error}
\item{pred_period}
}
\description{
Estimate the predictive error of an index via an in-sample approach. Generic method.
}
\section{Further Details}{

In addition to being a stand-alone function, it is also used by `calcForecastError` and `calcKFoldError``
}

\examples{
\dontrun{
index_error <- calcInSampleError(pred_df = rt_index$data,
                                 index = rt_index$index$index)
}
}
