% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcKFoldError.R
\name{calcKFoldError}
\alias{calcKFoldError}
\title{calcKFoldError}
\usage{
calcKFoldError(hpi_obj, pred_df, k, seed)
}
\arguments{
\item{hpi_obj}{Hpi object of class 'hpi'}

\item{pred_df}{Data.frame of sales to be used for assessing predictive quality of index}

\item{k}{default=10; Number of folds to apply to holdout process}

\item{seed}{default=1; Random seed generator to control the folding process}

\item{smooth}{default = FALSE; Calculate on the smoothed index}

\item{...}{Additional Arguments}
}
\value{
object of class `hpiaccuracy` inheriting from class `data.frame` containing the following fields:
\describe{
\item{prop_id}{Property Identification number}
\item{pred_price}{Predicted price}
\item{pred_error}{(Prediction - Actual) / Actual}
\item{pred_period}{Period of the prediction}
}
}
\description{
Estimate out-of-sample index errors using a KFold process
}
\examples{
# Load data
  data(ex_hpi)
  data(ex_rtdata)

# Calc Accuracy
  kf_accr <- calcKFoldError(hpi_obj = ex_hpi,
                            pred_df = ex_rtdata,
                            k = 10,
                            seed = 123,
                            smooth = FALSE)

}
