% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcKFoldErrors.R
\name{calcKFoldErrors}
\alias{calcKFoldErrors}
\title{calcKFoldErrors}
\usage{
Lorem Ipsum...
}
\arguments{
\item{hpi_obj}{Object of class 'hpi'}

\item{pred_data}{Set of sales to be used for predicitive quality of index}

\item{k}{default=10; Number of folds to apply to holdout process}

\item{seed}{random seed generator to control the folding process}

\item{...}{Additional Arguments}
}
\value{
hpimodel object
}
\description{
Estimate out-of-sample index errors using a KFold process
}
\section{Further Details}{

Lorem Ipsum...
}

\examples{
a <- 1
}
