% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAccuracy.R
\name{calcSeriesAccuracy}
\alias{calcSeriesAccuracy}
\title{calcSeriesAccuracy}
\usage{
Lorem Ipsum...
}
\arguments{
\item{series_obj}{Series object to be calculted}

\item{test_method}{default = 'insample'; also 'kfold' or 'forecast'}

\item{test_type}{default = 'rt'; Type of data to use for test.  See details.}

\item{pred_df}{= NULL; Extra data if the test_type doesn't match data in hpi_obj}

\item{...}{Additional Arguments}
}
\value{
`serieshpi` object
}
\description{
Calculates accuracy over a series of indexes
}
\section{Further Details}{

Leaving order blank default to a moving average with order 3.
}

