% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcVolatility.R
\name{calcSeriesVolatility}
\alias{calcSeriesVolatility}
\title{calcSeriesVolatility}
\usage{
Lorem Ipsum...
}
\arguments{
\item{series_obj}{Series object to be calculted}

\item{window}{default = 3; Rolling periods over which to calculate the volatility}

\item{smooth}{default = FALSE; Also calculate volatilities for smoothed indexes}

\item{...}{Additional Arguments}
}
\value{
`serieshpi` object
}
\description{
Calculates volatility over a series of indexes
}
\section{Further Details}{

Leaving order blank default to a moving average with order 3.
}

