% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcVolatility.R
\name{calcVolatility}
\alias{calcVolatility}
\title{calcVolatility}
\usage{
calcVolatility(index, window, in_place = FALSE,
  in_place_name = "volatility", ...)
}
\arguments{
\item{index}{time-series index object}

\item{window}{periods over which to calculate the volatility}

\item{in_place}{default = FALSE; adds volatility metric to the `hpiindex` object
(may be within an `hpi` object)}

\item{in_place_name}{default = 'vol'; name of volatility object in `hpiindex` object}

\item{...}{Additional arguments}
}
\value{
A metric of volatility
}
\description{
Calculate index volatility
}
\section{Further Details}{

}

\examples{
index_vol <- calcVolatility(index=index,
                            window=3)
}
