% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateToPeriod.R
\name{checkDate}
\alias{checkDate}
\title{checkDate}
\usage{
checkDate(x_date, name)
}
\arguments{
\item{x_date}{Date string or vector}

\item{name}{Name of argument to return in error/warning message}
}
\value{
Adjusted date field
}
\description{
Checks and converts date arguments into proper format
}
