% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcKFoldError.R
\name{createKFoldData}
\alias{createKFoldData}
\title{createKFoldData}
\usage{
Lorem Ipsum...
}
\arguments{
\item{score_ids}{Vector of row ids to be included in scoring data}

\item{full_data}{Complete dataset of this model type}

\item{pred_df}{Data to be used for prediction}
}
\value{
list
\item{train} Training data
\item{score} Scoring data
}
\description{
create the datasets for the kfold error testing (Generic Method)
}
\section{Further Details}{

Lorem Ipsum...
}

\examples{
a <- 1
}
