% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcKFoldError.R
\name{createKFoldData.rtdata}
\alias{createKFoldData.rtdata}
\title{Create data for KFold error test (rt approach)}
\usage{
\method{createKFoldData}{rtdata}(score_ids, full_data, pred_df)
}
\arguments{
\item{score_ids}{Vector of row ids to be included in scoring data}

\item{full_data}{Complete dataset (class `hpidata``) of this model type (rt or hed)}

\item{pred_df}{Data to be used for prediction}
}
\description{
`rtdata` method for creating KFold testing data
}
