% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpir_data.R
\docType{data}
\name{ex_heddata}
\alias{ex_heddata}
\title{Example `heddata` object}
\format{A \code{"heddata"}, \code{"hpidata"} and \code{"data.frame"} with 5,319 rows
and 16 variables
\describe{
  \item{prop_id}{The unique property identifying code.  Original value is preceded by
  two '..'s to prevent the dropping of leading zeros}
  \item{trans_id}{The unique transaction identifying code.}
  \item{price}{Price of the home}
  \item{sale_date}{Date of sale}
  \item{use_type}{Property use type}
  \item{area}{Assessment area or zone}
  \item{lot_sf}{Size of lot in square feet}
  \item{wfnt}{Is property waterfront?}
  \item{bldg_grade}{Quality of the building construction (higher is better)}
  \item{tot_sf}{Size of home in square feet}
  \item{beds}{Number of bedrooms}
  \item{baths}{Number of bathrooms}
  \item{age}{Age of home}
  \item{eff_age}{Age of home, considering major remodels}
  \item{longitude}{Longitude}
  \item{latitude}{Latitude }
  \item{trans_date}{Formatted R date field}
  \item{trans_period}{Standardized relative period value (integer)}
}}
\source{
Output from the hedCreateTrans() function
}
\usage{
data(ex_heddata)
}
\description{
Based off of the `ex_hpidata` object.  This is an `heddata` object as indicated by the
presence of the five standardize fields: "prop_id", "trans_id", "price", "trans_date"
and "trans_period".  All other fields can vary based on the raw data being used
}
\keyword{datasets}
