% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpir_data.R
\docType{data}
\name{ex_hedmodel}
\alias{ex_hedmodel}
\title{Example `hedmodel` object}
\format{List of 12 objects (`lm` object)
\describe{
  \item{coefficients}{Vector of coefficients, length 87}
  \item{residuals}{Vector of residuals, length 5319}
  \item{effects}{Model effects, length 5319}
  \item{rank}{model rank, length 87}
  \item{fitted.values}{Fitted values, length 5319}
  \item{assign}{Vector of length 87}
  \item{df.residual}{Residual degrees of freedom, integer}
  \item{contrasts}{list of contrasts}
  \item{xlevels}{levels}
  \item{call}{model call}
  \item{terms}{formula terms of the model}
  \item{model}{model dep and ind variables}
}}
\source{
hedModel()
}
\usage{
data(ex_hedmodel)
}
\description{
An object of class `hedmodel`, resulting from `hedModel()`
}
\keyword{datasets}
