% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpir_data.R
\docType{data}
\name{ex_hpi}
\alias{ex_hpi}
\title{Example `hpi` model}
\format{List of three objects
\describe{
  \item{data}{Object of class `hpidata` and `rtdata` or `heddata`}
  \item{model}{Object of class `hpimodel`}
  \item{index}{Object of class `hpiindex`}
}}
\source{
`rtIndex()` or `hedIndex()` functions
}
\usage{
data(ex_hpi)
}
\description{
An example hpi collection object of class `hpi`
}
\keyword{datasets}
