% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpir_data.R
\docType{data}
\name{ex_hpiaccuracy}
\alias{ex_hpiaccuracy}
\title{Example `hpiaccuracy` object}
\format{`data.frame` containing four variables:
\describe{
  \item{prop_id}{Unique property ID}
  \item{pred_price}{Predicted price}
  \item{pred_error}{Error of the prediction}
  \item{pred_period}{Prediction period}
}}
\source{
From the `calcAccuracy()` function`
}
\usage{
data(ex_hpiaccuracy)
}
\description{
Object containing accuracy information for an `hpi` object
}
\keyword{datasets}
