% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpir_data.R
\docType{data}
\name{ex_hpiindex}
\alias{ex_hpiindex}
\title{Example `hpiindex` object}
\format{List containing five objects:
\describe{
  \item{name}{Vector of period names, length 84}
  \item{numeric}{Vector of numeric period values, length 84}
  \item{period}{Vector of period numbers, length 84}
  \item{value}{Vector of index values, length 84}
  \item{imputed}{Vector showing if value was imputed, length 84}
}}
\source{
`modelToIndex()` function
}
\usage{
data(ex_hpiindex)
}
\description{
An object of class `hpiindex`
}
\keyword{datasets}
