% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpir_data.R
\docType{data}
\name{ex_hpimodel}
\alias{ex_hpimodel}
\title{Example `hpimodel` object}
\format{List containing 8 objects:
\describe{
  \item{estimator}{type of estimator used}
  \item{coefficients}{data.frame of model coefficients}
  \item{model_obj}{`rtmodel` object}
  \item{mod_spec}{Full Model specification}
  \item{log_dep}{Is the dependent variable in log format?}
  \item{base_price}{mean price in period 1}
  \item{periods}{data.frame of period data (object)}
  \item{approach}{Approach used (rt or hed)}
}}
\source{
`hpiModel()``
}
\usage{
data(ex_hpimodel)
}
\description{
An `hpimodel` object example (from an `rtModel` or `hedModel` object)
}
\keyword{datasets}
