% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpir_data.R
\docType{data}
\name{ex_indexvolatility}
\alias{ex_indexvolatility}
\title{Example `indexvolatility` object}
\format{List of length three containing:
\describe{
  \item{roll}{vector of period volatilities, length 79}
  \item{mean}{mean of volatility}
  \item{median}{median of volatility}
}}
\source{
`calcVolatility()`
}
\usage{
data(ex_indexvolatility)
}
\description{
An example of an `indexvolatility` object
}
\keyword{datasets}
