% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpir_data.R
\docType{data}
\name{ex_plotaccuracy}
\alias{ex_plotaccuracy}
\title{Example `plotaccuracy` object}
\format{ggplot object:}
\source{
plot.hpiaccuracy()
}
\usage{
data(ex_plotaccuracy)
}
\description{
An example plotaccuracy object
}
\keyword{datasets}
