% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpir_data.R
\docType{data}
\name{ex_plotindex}
\alias{ex_plotindex}
\title{Example `plotindex` object}
\format{ggplot object:}
\source{
plot.hpiindex() or plot.hpi()
}
\usage{
data(ex_plotindex)
}
\description{
An example plotindex object
}
\keyword{datasets}
