% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpir_data.R
\docType{data}
\name{ex_plotrevision}
\alias{ex_plotrevision}
\title{Example `plotrevision` object}
\format{ggplot object:}
\source{
plot.seriesrevision()
}
\usage{
data(ex_plotrevision)
}
\description{
An example plotrevision object
}
\keyword{datasets}
