% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpir_data.R
\docType{data}
\name{ex_plotseries}
\alias{ex_plotseries}
\title{Example `plotseries` object}
\format{ggplot object:}
\source{
plot.serieshpi()
}
\usage{
data(ex_plotseries)
}
\description{
An example plotseries object
}
\keyword{datasets}
