% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpir_data.R
\docType{data}
\name{ex_plotvolatility}
\alias{ex_plotvolatility}
\title{Example `plotvolatility` object}
\format{ggplot object:}
\source{
plot.indexvolatility()
}
\usage{
data(ex_plotvolatility)
}
\description{
An example plotvolatility object
}
\keyword{datasets}
