% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpir_data.R
\docType{data}
\name{ex_rtdata}
\alias{ex_rtdata}
\title{Example `rtdata` object}
\format{A \code{"rtdata"}, \code{"hpidata"} and \code{"data.frame"} with 656 rows
and 8 variables
\describe{
  \item{prop_id}{The unique property identifying code}
  \item{period_1}{Period of the first transaction}
  \item{period_2}{Period of the second transaction}
  \item{price_1}{First Price}
  \item{price_2}{Second Price}
  \item{trans_id1}{First Unique transaction ID}
  \item{trans_id2}{Second Unique transaction ID}
  \item{pair_id}{Unique pairing ID}
}}
\source{
Output from the rtCreateTrans() function
}
\usage{
data(ex_rtdata)
}
\description{
Based off of the `ex_hpidata` object.  This is an `rtdata` object as indicated by the
standardized nature of the object.  All `rtdata` object contain the same eight fields.
}
\keyword{datasets}
