% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpir_data.R
\docType{data}
\name{ex_rtmodel}
\alias{ex_rtmodel}
\title{Example `rtmodel` object}
\format{List of 11 objects (`lm` object)
\describe{
  \item{coefficients}{Vector of coefficients, length 87}
  \item{residuals}{Vector of residuals, length 5319}
  \item{effects}{Model effects, length 5319}
  \item{rank}{model rank, length 87}
  \item{fitted.values}{Fitted values, length 5319}
  \item{assign}{Vector of length 87}
  \item{df.residual}{Residual degrees of freeedom, integer}
  \item{xlevels}{levels}
  \item{call}{model call}
  \item{terms}{formula terms of the model}
  \item{model}{model dep and ind variables}
}}
\source{
rtModel()
}
\usage{
data(ex_rtmodel)
}
\description{
An object of class `rtmodel`, resulting from `rtModel()`
}
\keyword{datasets}
