% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpir_data.R
\docType{data}
\name{ex_seriesaccuracy}
\alias{ex_seriesaccuracy}
\title{Example `seriesaccuracy` object}
\format{`data.frame` containing five variables:
\describe{
  \item{prop_id}{Unique property ID}
  \item{pred_price}{Predicted price}
  \item{pred_error}{Error of the prediction}
  \item{pred_period}{Prediction period}
  \item{series}{series number from which prediction was generated}
}}
\source{
From the `calcSeriesAccuracy()` function`
}
\usage{
data(ex_seriesaccuracy)
}
\description{
Object containing accuracy information for a `serieshpi` object
}
\keyword{datasets}
