% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpir_data.R
\docType{data}
\name{ex_serieshpi}
\alias{ex_serieshpi}
\title{Example `serieshpi` object}
\format{list of length 2 containing
\describe{
  \item{data}{series data (class `rtdata`)}
  \item{hpis}{list of hpi objects (length 13)}
}}
\source{
`createSeries()`
}
\usage{
data(ex_serieshpi)
}
\description{
An example `serieshpi` object
}
\keyword{datasets}
