% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpir_data.R
\docType{data}
\name{ex_seriesrevision}
\alias{ex_seriesrevision}
\title{Example `seriesrevision` object}
\format{List of 3 containing:
\describe{
  \item{period}{data.frame with period number, mean and median revision for each period}
  \item{median}{median revision of all periods}
  \item{mean}{mean revision of all periods}
}}
\source{
`calcRevision()``
}
\usage{
data(ex_seriesrevision)
}
\description{
Revision statistics for a series
}
\keyword{datasets}
