% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpir_data.R
\docType{data}
\name{ex_smoothindex}
\alias{ex_smoothindex}
\title{Example `smoothindex` object}
\format{`ts` object of length 84}
\source{
`smoothIndex()`
}
\usage{
data(ex_smoothindex)
}
\description{
An index that has been smoothed
}
\keyword{datasets}
