% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpir_data.R
\docType{data}
\name{ex_timematrix}
\alias{ex_timematrix}
\title{Example `timematrix` object}
\format{array of 656 rows and 83 columns}
\source{
`rtTimeMatrix()``
}
\usage{
data(ex_timematrix)
}
\description{
Time matrix object that is used in the `rtmodel()` (repeat transaction approach)
}
\keyword{datasets}
