% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hedCreateSales.R
\name{hedCreateSales}
\alias{hedCreateSales}
\title{hedCreateSales}
\usage{
hedCreateSales(sales_df, prop_id, sale_id, price, date = NULL,
  periodicity = NULL, ...)
}
\arguments{
\item{sales_df}{sales transaction in either a data.frame or a
sales.df class from dateToPeriod() function}

\item{prop_id}{field contain the unique property identification}

\item{sale_id}{field containing the unique sale identification}

\item{price}{field containing the sale price}

\item{date}{default=NULL, field containing the date of the sale.
Only necessary if not passing a sales.df object}

\item{periodicity}{default=NULL, field containing the desired periodicity of analysis.
Only necessary if not passing a sales.df object}
}
\value{
data.frame of repeat sales. Note that a full data.frame of the possible
periods, their values and names can be found in the attributes to the returned `hed` object
}
\description{
Create a hedonic model ready data.frame of sale transactions
}
\section{Further Details}{

aaa
}

\examples{
## With a raw data.frame
hed_sales <- hedCreateSales(sales_df = seattle_sales,
                           prop_id = 'pinx',
                           sale_id = 'uniq_id',
                           price = 'sale_price',
                           date = 'sale_date',
                           periodicity = 'monthly')
}
