% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hedModel.R
\name{hedModel.base}
\alias{hedModel.base}
\title{hedModel.base}
\usage{
\method{hedModel}{base}(estimator, hed_df, hed_spec, ...)
}
\arguments{
\item{estimator}{Type of model to estimates (base, robust, weighted)}

\item{hed_df}{Repeat sales dataset from rsCreateSales()}

\item{hed_spec}{twen the two sales}

\item{...}{Additional arguments}
}
\value{
rs model object
}
\description{
Estimate repeat sales model (method based on estimator class). Generic method.
}
\section{Further Details}{

}

\examples{
rs_model <- rsModel(estimator = 'base',
                    rs_df = rs_sales,
                    time_matrix = time_matrix,
                    price_diff = price_diff)
}
