% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hedModel.R
\name{hedModel}
\alias{hedModel}
\title{hedModel}
\usage{
hedModel(estimator, hed_df, hed_spec, ...)
}
\arguments{
\item{estimator}{Type of model to estimates (base, robust, weighted)}

\item{hed_df}{Repeat sales dataset from rtCreateSales()}

\item{hed_spec}{Time matrix object from rtTimeMatrix()}

\item{...}{Additional arguments}

\item{price_diff}{Difference in price betwen the two sales}
}
\value{
`hedmodel` object: model object of the estimator (ex.: `lm`)
}
\description{
Estimate repeat sales model (method based on estimator class). Generic method.
}
\section{Further Details}{

`estimator` argument must be in a class of 'base', 'weighted' or 'robust'

This function is not generally called directly, but rather from `hpiModel()`
}

\examples{
# Load Data
data(ex_heddata)

# Estimate Model
hed_model <- hedModel(estimator = structure('base', class = 'base'),
                      hed_df = ex_heddata,
                      hed_spec = as.formula(log(price) ~ baths + tot_sf))
}
