% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hedModel.R
\name{hedModel.weighted}
\alias{hedModel.weighted}
\title{hedModel.weighted}
\usage{
\method{hedModel}{weighted}(estimator, hed_df, hed_spec, ...)
}
\arguments{
\item{estimator}{Type of model to estimates (base, robust, weighted)}

\item{hed_df}{Repeat sales dataset from hedCreateSales()}

\item{hed_spec}{Model specfication}

\item{...}{Additional arguments}
}
\value{
hed model object
}
\description{
Estimate hedonic model (method based on estimator class). Generic method.
}
\section{Further Details}{

}

