% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpiModel.R
\name{hpiModel.heddata}
\alias{hpiModel.heddata}
\title{Specific method for hpi modeling (hed) approach)}
\usage{
\method{hpiModel}{heddata}(hpi_df, estimator = "base", log_dep = TRUE,
  trim_model = TRUE, hed_spec = NULL, dep_var = NULL, ind_var = NULL,
  ...)
}
\arguments{
\item{hpi_df}{Dataset created by one of the *CreateSales() function in this package.}

\item{estimator}{Type of estimator to be used ('base', 'weighted', 'robust')}

\item{log_dep}{default=TRUE; should the dependent variable (change in price) be logged?}

\item{trim_model}{default TRUE, should excess be trimmed from model results ('lm' or 'rlm' object)?}

\item{hed_spec}{default=NULL; hedonic model specification}

\item{dep_var}{default=NULL; dependent variable of the model}

\item{ind_var}{default=NULL; independent variable(s) of the model}

\item{...}{Additional Arguments}
}
\value{
hpimodel object consisting of:
\describe{
\item{estimator}{Type of estimator}
\item{coefficients}{Data.frame of coefficient}
\item{model_obj}{class `rtmodel` or `hedmodel`}
\item{mod_spec}{Full model specification}
\item{log_dep}{Binary: is the dependent variable in logged format}
\item{base_price}{Mean price in the base period}
\item{periods}{`data.frame` of periods}
\item{approach}{Type of model used}
}
}
\description{
Estimate hpi models with hed approach
}
\examples{
# Load Data
 data(ex_heddata)

 # Create model object
 hpi_model <- hpiModel(hpi_df = ex_heddata,
                       estimator = 'base',
                       dep_var = 'price',
                       ind_var = c('tot_sf', 'beds', 'baths'),
                       log_dep = TRUE)
}
