% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpiModel.R
\name{hpiModel}
\alias{hpiModel}
\title{hpiModel}
\usage{
Lorem Ipsum...
}
\arguments{
\item{hpi_df}{Dataset created by one of the *CreateTrans() function in this package.}

\item{estimator}{Type of estimator to be used ('base', 'weighted', 'robust')}

\item{log_dep}{default TRUE, should the dependent variable (change in price) be logged?}

\item{trim_model}{default TRUE, should excess be trimmed from model results ('lm' or 'rlm' object)?}

\item{...}{Additional Arguments}
}
\value{
hpimodel object
}
\description{
Estimate the model for any method of house price index.  Generic method.
}
\section{Further Details}{

Lorem Ipsum...
}

\examples{
a <- 1
}
