% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpiModel.R
\name{hpiModel.rs}
\alias{hpiModel.rs}
\title{hpiModel.rs}
\usage{
Lorem Ipsum...
}
\arguments{
\item{hpi_data}{Dataset created by one of the *CreateSales() function in this package.}

\item{estimator}{Type of estimator to be used ('base', 'weighted', 'robust')}

\item{log_dep}{default TRUE, should the dependent variable (change in price) be logged?}

\item{...}{Additional Arguments}
}
\value{
hpimodel object
}
\description{
Estimate the model for any method of house price index.  Generic method.
}
\section{Further Details}{

Lorem Ipsum...
}

\examples{
sea_sales <- dateToPeriod(sales_df = seattle_sales,
                          date = 'sale_date',
                          periodicity = 'month')
rep_sales <- rsCreateSales(sales_df = sea_sales,
                           prop_id = 'pinx',
                           sale_id = 'uniq_id',
                           price = 'sale_price')
rs_model <- hpiModel(hpi_data = rep_sales,
                     estimator = 'base',
                     log_dep = TRUE)
}
