% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpiModel.R
\name{hpiModel.rtdata}
\alias{hpiModel.rtdata}
\title{hpiModel.rtdata}
\usage{
Lorem Ipsum...
}
\arguments{
\item{hpi_df}{Dataset created by one of the *CreateTrans() function in this package.}

\item{estimator}{Type of estimator to be used ('base', 'weighted', 'robust')}

\item{log_dep}{default TRUE, should the dependent variable (change in price) be logged?}

\item{trim_model}{default TRUE, should excess be trimmed from model results ('lm' or 'rlm' object)?}

\item{...}{Additional Arguments}
}
\value{
hpimodel object
}
\description{
Estimate the model for any method of house price index.  Generic method.
}
\section{Further Details}{

Lorem Ipsum...
}

\examples{
sea_sales <- dateToPeriod(trans_df = seattle_sales,
                          date = 'sale_date',
                          periodicity = 'month')
rep_sales <- rtCreateTrans(trans_df = sea_sales,
                           prop_id = 'pinx',
                           trans_id = 'uniq_id',
                           price = 'sale_price')
rt_model <- hpiModel(hpi_df = rep_sales,
                     estimator = 'base',
                     log_dep = TRUE)
}
