% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpiR.R
\docType{package}
\name{hpiR}
\alias{hpiR}
\alias{hpiR-package}
\title{hpiR: A package for house price indexes}
\description{
House Price Indexes in R: A set of tools to create house price indexes and analyze
their various performance metrics.
}
