% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeSeries.R
\name{imputeSeries}
\alias{imputeSeries}
\title{imputeSeries}
\usage{
imputeSeries(coef_df, max_period = max(coef_df$time), ...)
}
\arguments{
\item{coef_df}{data.frame of coefficient values from hpiModel() function}

\item{max_period}{Maximum number of periods that should have been estimated.}

\item{...}{Additional arguments}
}
\value{
data.frame of coefficient values
}
\description{
imputes missing time periods in the results of an hpimodel object
}
\section{Further Details}{

}

\examples{
hpi_coef <- imputeSeries(coef_df,
                          max_period=84)
}
