% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcForecastErrors.R
\name{makeFCData}
\alias{makeFCData}
\title{makeFCData}
\usage{
Lorem Ipsum...
}
\arguments{
\item{time_cut}{Period of forecast}

\item{hpi_data}{Data to be converted to training or scoring}

\item{train}{Default=TRUE; Create training data?  FALSE = Scoring data}

\item{...}{Additional Arguments}
}
\value{
data.frame of training or scoring observations
}
\description{
Create training or scoring data for the forecast error calculations
}
\section{Further Details}{

Lorem Ispum...
}

\examples{
a <- 1
}
