% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcKFoldError.R
\name{matchKFold.heddata}
\alias{matchKFold.heddata}
\title{Helper function to make KFold data}
\usage{
\method{matchKFold}{heddata}(train_df, pred_df)
}
\arguments{
\item{train_df}{Data.frame of training data}

\item{pred_df}{Data.frame (class `hpidata``) to be used for prediction}
}
\description{
Function to help create KFold data based on hed approach
}
