% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcKFoldError.R
\name{matchKFold}
\alias{matchKFold}
\title{matchKFoldData}
\usage{
matchKFold(train_df, pred_df)
}
\arguments{
\item{train_df}{Data.frame of training data}

\item{pred_df}{Data.frame (class `hpi_df``) to be used for prediction}
}
\value{
list
\item{train} Training data
\item{score} Scoring data
}
\description{
Makes specific selections of scoring data (Generic Method)
}
\section{Further Details}{

Called from createKFoldData
}

\examples{
\dontrun{
kfold_df$score <- matchKFold(train_df = kfold_df$train,
                             pred_df = hpi_obj$pred_data)
}
}
