% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelToIndex.R
\name{modelToIndex}
\alias{modelToIndex}
\title{modelToIndex}
\usage{
modelToIndex(hpimodel, max_period = max(hpimodel$coefficients$time), ...)
}
\arguments{
\item{hpimodel}{Model results object}

\item{max_period}{Maximum number of periods that should have been estimated.}

\item{...}{Additional arguments}
}
\value{
rs model object
}
\description{
Converts model results into a house price index
}
\section{Further Details}{

}

\examples{
hpi_index <- modelToIndex(hpi_model,
                          max_period=84)
}
