% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateToPeriod.R
\name{periodTable.annual}
\alias{periodTable.annual}
\title{Create a table of the annual periods}
\usage{
\method{periodTable}{annual}(trans_df, periodicity)
}
\arguments{
\item{trans_df}{Transaction data.frame}

\item{periodicity}{Periodicity option ('weekly', 'monthly', 'quarterly', 'annually')}
}
\description{
Specific method for creating annual period table
}
