% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateToPeriod.R
\name{periodTable.equalfreq}
\alias{periodTable.equalfreq}
\title{Create a table of equal frequency (any frequency) periods}
\usage{
\method{periodTable}{equalfreq}(trans_df, periodicity, freq = NULL,
  start = NULL, first_date = NULL, last_date = NULL, ...)
}
\arguments{
\item{trans_df}{Transaction data.frame}

\item{periodicity}{Periodicity option ('weekly', 'monthly', 'quarterly', 'annually')}

\item{freq}{[30] Frequency width of each period in days}

\item{start}{['first'] Where to start counting ('first' or 'last')}

\item{first_date}{[NULL] If null, the data determines the first date.  Else set your own.
Note that the first_date must be outside of the range of transaction dates.  It can only extend
the time period, not clip it.  That should be done else where.}

\item{last_date}{[NULL] If null, the data determines the last date. Else set your own
Note that the last_date must be outside of the range of transaction dates.  It can only extend
the time period, not clip it.  That should be done else where.}

\item{...}{Additional Arguments}
}
\description{
Specific method for creating flexible frequency periods
}
