% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateToPeriod.R
\name{periodTable.equalsample}
\alias{periodTable.equalsample}
\title{Create a table of equal frequency (any frequency) periods}
\usage{
\method{periodTable}{equalsample}(trans_df, periodicity, nbr_periods, ...)
}
\arguments{
\item{trans_df}{Transaction data.frame}

\item{periodicity}{Periodicity option ('weekly', 'monthly', 'quarterly', 'annually')}

\item{nbr_periods}{Number of periods to use}

\item{...}{Additional Arguments}
}
\description{
Specific method for creating flexible frequency periods
}
