% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateToPeriod.R
\name{periodTable.monthly}
\alias{periodTable.monthly}
\title{Create a table of the monthly periods}
\usage{
\method{periodTable}{monthly}(trans_df, periodicity)
}
\arguments{
\item{trans_df}{Transaction data.frame}

\item{periodicity}{Periodicity option ('weekly', 'monthly', 'quarterly', 'annually')}
}
\description{
Specific method for creating monthly period table
}
