% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateToPeriod.R
\name{periodTable.quarterly}
\alias{periodTable.quarterly}
\title{Create a table of the quarterly periods}
\usage{
\method{periodTable}{quarterly}(trans_df, periodicity)
}
\arguments{
\item{trans_df}{Transaction data.frame}

\item{periodicity}{Periodicity option ('weekly', 'monthly', 'quarterly', 'annually')}
}
\description{
Specific method for creating quarterly period table
}
