% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateToPeriod.R
\name{periodTable}
\alias{periodTable}
\title{Create a table of the periods (generic method)}
\usage{
periodTable(trans_df, periodicity)
}
\arguments{
\item{trans_df}{Transaction data.frame}

\item{periodicity}{Periodicity option ('weekly', 'monthly', 'quarterly', 'annually')}

\item{...}{Additional Arguments}
}
\value{
[data.frame] consisting of
\describe{
\item{period}{Period number}
\item{start_date}{start date of each period}
\item{end_date}{end date of each period}
\item{name}{name of the period}
}
}
\description{
Generic method for create simple table of all selected periods.  Used within `dateToPeriod()`
}
\examples{

 # Load data
 data(ex_sales)
 ex_sales$trans_date <- checkDate(ex_sales[['sale_date']], 'date')

 # With a raw transaction data.frame
 pt_df <- periodTable(trans_df = ex_sales,
                      periodicity = 'annual')
}
