% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateToPeriod.R
\name{periodTable.weekly}
\alias{periodTable.weekly}
\title{Create a table of the weekly periods}
\usage{
\method{periodTable}{weekly}(trans_df, periodicity)
}
\arguments{
\item{trans_df}{Transaction data.frame}

\item{periodicity}{Periodicity option ('weekly', 'monthly', 'quarterly', 'annually')}
}
\description{
Specific method for creating weekly period table
}
