% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plot.hpiaccuracy}
\alias{plot.hpiaccuracy}
\title{plot.hpiaccuracy}
\usage{
plot(x=accr_obj, return_plot, do_plot)
}
\arguments{
\item{x}{Object to plot of class `hpiaccuracy``}

\item{return_plot}{default = FALSE; Return the plot to the function call}

\item{do_plot}{default = FALSE; Execute plotting to terminal/console}
}
\value{
`plotaccuracy` object inheriting from a ggplot object
}
\description{
Simple Plot of an hpiaccuracy object
}
\examples{
# Load data
 data(ex_hpiaccuracy)

# Make Plot
 plot(ex_hpiaccuracy)
}
