% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plot.hpiindex}
\alias{plot.hpiindex}
\title{Plot method for `hpiindex` object}
\usage{
\method{plot}{hpiindex}(x, show_imputed = FALSE, smooth = FALSE, ...)
}
\arguments{
\item{x}{Object to plot of class `hpiindex``}

\item{show_imputed}{default = FALSE; highlight the imputed points}

\item{smooth}{default = FALSE; plot the smoothed index}

\item{...}{Additional Arguments}
}
\value{
`plotindex` object inheriting from a ggplot object
}
\description{
Specific plotting method for hpiindex objects
}
\examples{
# Load data
 data(ex_hpiindex)

# Make Plot
 plot(ex_hpiindex)
}
