% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plot.indexvolatility}
\alias{plot.indexvolatility}
\title{Plot method for `indexvolatility` object}
\usage{
\method{plot}{indexvolatility}(x, ...)
}
\arguments{
\item{x}{Object to plot of class `indexvolatility``}

\item{...}{Additional Arguments}
}
\value{
`plotvolatility` object inheriting from a ggplot object
}
\description{
Specific plotting method for indexvolatility objects
}
\examples{
# Load data
 data(ex_indexvolatility)

# Make Plot
 plot(ex_indexvolatility)
}
