% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plot.seriesaccuracy}
\alias{plot.seriesaccuracy}
\title{Plot method for `seriesaccuracy` object}
\usage{
\method{plot}{seriesaccuracy}(x, return_plot = FALSE, ...)
}
\arguments{
\item{x}{Object of class `hpiaccuracy``}

\item{return_plot}{default = FALSE; Return the plot to the function call}

\item{...}{Additional argument (passed to `plot.hpiaccuracy()``)}
}
\value{
`plotaccuracy` object inheriting from a ggplot object
}
\description{
Specific plotting method for seriesaccuracy objects
}
\examples{
# Load data
 data(ex_seriesaccuracy)

# Make Plot
 plot(ex_seriesaccuracy)
}
