% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plot.serieshpi}
\alias{plot.serieshpi}
\title{Plot method for `serieshpi` object}
\usage{
\method{plot}{serieshpi}(x, smooth = FALSE, ...)
}
\arguments{
\item{x}{Object of class `serieshpi`}

\item{smooth}{default = FALSE; plot the smoothed object}

\item{...}{Additional Arguments`}
}
\value{
`plotseries` object inheriting from a ggplot object
}
\description{
Specific plotting method for serieshpi objects
}
\examples{
# Load data
 data(ex_serieshpi)

# Make Plot
 plot(ex_serieshpi)
}
