% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plot.seriesrevision}
\alias{plot.seriesrevision}
\title{plot.seriesrevision}
\usage{
plot(rev_obj)
}
\arguments{
\item{rev_obj}{Object of class `seriesrevision`}
}
\value{
`plotrevision` object inheriting from a ggplot object
}
\examples{
# Load data
 data(ex_seriesrevision)

# Make Plot
 plot(ex_seriesrevision)
}
