% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfModel.R
\name{rfModel.pdp}
\alias{rfModel.pdp}
\title{Random forest model approach with pdp estimator}
\usage{
\method{rfModel}{pdp}(estimator, rf_df, rf_spec, ntrees = 200,
  seed = 1, ...)
}
\arguments{
\item{estimator}{Type of model to estimates (pdp)}

\item{rf_df}{Transactions dataset from hedCreateSales()}

\item{rf_spec}{Model specification (`formula` object)}

\item{ntrees}{[200] Set number of trees to use}

\item{seed}{[1] Random seed for reproducibility}

\item{...}{Additional arguments}
}
\description{
Use of pdp estimator in random forest approach
}
\section{Further Details}{

See `?rfModel` for more information
}

