% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfModel.R
\name{rfModel.shap}
\alias{rfModel.shap}
\title{Hedonic model approach with base estimator}
\usage{
\method{rfModel}{shap}(estimator, rf_df, rf_spec, ntrees = 200,
  seed = 1, shap_k = 10, ...)
}
\arguments{
\item{estimator}{Type of model to estimates (base, robust, weighted)}

\item{rf_df}{Transactions dataset from hedCreateSales()}

\item{rf_spec}{Model specification (`formula` object)}

\item{ntrees}{[200] Set number of trees to use}

\item{seed}{[1] Random seed for reproducibility}

\item{...}{Additional arguments}
}
\description{
Use of base estimator in hedonic model approach
}
\section{Further Details}{

See `?rfModel` for more information
}

