% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfModel.R
\name{rfSim}
\alias{rfSim}
\title{Simulation engine}
\usage{
rfSim(rf_obj, sim_df, periods, ...)
}
\arguments{
\item{rf_obj}{A `ranger` random forest object}

\item{sim_df}{Single property to simulate over time}

\item{periods}{Time periods to simulate over}

\item{...}{Additional arguments}
}
\description{
Helper function to simulate each example proeprty over the time period(s)
'
}
