% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfModel.R
\name{rfSimDf}
\alias{rfSimDf}
\title{Create simulation data for Random forest approach}
\usage{
rfSimDf(rf_df, seed, ...)
}
\arguments{
\item{rf_df}{Full training dataset}

\item{seed}{Random seed for reproducibility}

\item{...}{Additional arguments}
}
\description{
Create data to use in PDP simulation
}
\section{Further Details}{

See `?rfModel` for more information
}

