% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfModel.R
\name{rfSimulate}
\alias{rfSimulate}
\title{Simulate selected properties}
\usage{
rfSimulate(rf_obj, rf_df, sim_type = "random", sim_per = 0.1,
  sim_count = NULL, ...)
}
\arguments{
\item{rf_obj}{A `ranger` random forest object}

\item{rf_df}{Full data.frame used to build the random forest}

\item{sim_type}{['random'] Sampling type to use}

\item{sim_per}{[0.1] Percentage of the total set to simulate}

\item{sim_count}{[NULL] If not giving a percentage, the total number of properties to simulate}

\item{...}{Additional arguments}
}
\description{
Handle simulation of all chosen properties
'
}
