% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsModel.R
\name{rsModel}
\alias{rsModel}
\title{rsModel}
\usage{
rsModel(rs_df, time_matrix, price_diff, estimator, ...)
}
\arguments{
\item{rs_df}{Repeat sales dataset from rsCreateSales()}

\item{time_matrix}{Time matrix object from rsTimeMatrix()}

\item{price_diff}{Difference in price betwen the two sales}

\item{estimator}{Type of model to estimates (base, robust, weighted).  Must be in that class.}

\item{...}{Additional arguments}
}
\value{
rs model object
}
\description{
Estimate repeat sales model (method based on estimator class). Generic method.
}
\section{Further Details}{

}

\examples{
rs_model <- rsModel(rs_df = rs_sales,
                    time_matrix = time_matrix,
                    price_diff = price_diff,
                    estimator = structure('base', class='base'))
}
