% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsModel.R
\name{rsModel.robust}
\alias{rsModel.robust}
\title{rsModel.robust}
\usage{
\method{rsModel}{robust}(rs_df, time_matrix, price_diff, estimator, ...)
}
\description{
Estimate repeat sales model (method based on estimator class). Generic method.
}
