% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsModel.R
\name{rsModel.weighted}
\alias{rsModel.weighted}
\title{rsModel.weighted}
\usage{
\method{rsModel}{weighted}(rs_df, time_matrix, price_diff, estimator, ...)
}
\description{
Estimate repeat sales model (method based on estimator class). Generic method.
}
