% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsTimeMatrix.R
\name{rsTimeMatrix}
\alias{rsTimeMatrix}
\title{rsTimeMatrix}
\usage{
rsTimeMatrix(rs_df)
}
\arguments{
\item{rs_df}{object of class `rs`: repeat sales data.frame created by rsCreateSales()}
}
\value{
matrix to be used on the right hand side of a repeat sales regression model
}
\description{
Create a time matrix for a repeat sale regression model
}
\section{Further Details}{

Time periods are calculated from the data provided.
}

\examples{
rep_sales <- rsCreateSales(sales_df = seattle_sales,
                           prop_id = 'pinx',
                           sale_id = 'uniq_id',
                           date = 'sale_date',
                           price = 'sale_price',
                           periodicity = 'qtr')
rs_matrix <- rsTimeMatrix(rep_sales)
}
