% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtModel.R
\name{rtModel.base}
\alias{rtModel.base}
\title{rtModel.base}
\usage{
\method{rtModel}{base}(rt_df, time_matrix, price_diff, estimator, ...)
}
\description{
Estimate repeat sales model (method based on estimator class). Generic method.
}
