% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtModel.R
\name{rtModel}
\alias{rtModel}
\title{rtModel}
\usage{
rtModel(rt_df, price_diff, time_matrix, estimator, ...)
}
\arguments{
\item{rt_df}{Repeat transactions dataset from rtCreateTrans()}

\item{time_matrix}{Time matrix object from rtTimeMatrix()}

\item{price_diff}{Difference in price betwen the two transactions}

\item{estimator}{Type of model to estimates (base, robust, weighted).  Must be in that class.}

\item{...}{Additional arguments}
}
\value{
`rtmodel` object
}
\description{
Estimate repeat transactions model (method based on estimator class). Generic method.
}
\section{Further Details}{

}

\examples{
 # Load Data
 data(ex_rtdata)
 data(ex_timematrix)

 # Calc price differences
 price_diff <- ex_rtdata$price_2 - ex_rtdata$price_1

 # Calculate model
 rt_model <- rtModel(rt_df = ex_rtdata,
                     price_diff = price_diff,
                     time_matrix = ex_timematrix,
                     estimator = structure('base', class='base'))
}
