% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtModel.R
\name{rtModel.robust}
\alias{rtModel.robust}
\title{rtModel.robust}
\usage{
\method{rtModel}{robust}(rt_df, time_matrix, price_diff, estimator, ...)
}
\description{
Estimate repeat sales model (method based on estimator class). Generic method.
}
