% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtTimeMatrix.R
\name{rtTimeMatrix}
\alias{rtTimeMatrix}
\title{rtTimeMatrix}
\usage{
rtTimeMatrix(rt_df)
}
\arguments{
\item{rt_df}{object of class `rt`: repeat transaction data.frame created by rtCreateTrans()}
}
\value{
matrix to be used on the right hand side of a repeat sales regression model
}
\description{
Create a time matrix for a repeat transaction regression model
}
\section{Further Details}{

Time periods are calculated from the data provided.
}

\examples{
rep_sales <- rtCreateTrans(trans_df = seattle_sales,
                           prop_id = 'pinx',
                           sale_id = 'uniq_id',
                           date = 'sale_date',
                           price = 'sale_price',
                           periodicity = 'qtr')
rt_matrix <- rtTimeMatrix(rep_sales)
}
