% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothIndex.R
\name{smoothIndex}
\alias{smoothIndex}
\title{Smooth an index}
\usage{
smoothIndex(index_obj, order = 3, in_place = FALSE, ...)
}
\arguments{
\item{index_obj}{Index to be smoothed}

\item{order}{default = 3; Number of nearby period to smooth with, multiple means
multiple iterations}

\item{in_place}{default = FALSE; adds smoothed index to the `hpiindex` object}

\item{...}{Additional Arguments}
}
\value{
a `ts`` and 'smooth_index` object with smoothed index
}
\description{
Smoothes an existing hpiindex object
}
\section{Further Details}{

Leaving order blank default to a moving average with order 3.
}

\examples{
# Load Data
data(ex_hpiindex)

# Create Smooth index
sm_index <- smoothIndex(index_obj = ex_hpiindex,
                        order = 3,
                        in_place = FALSE)

# Create Smooth index (in place)
sm_index <- smoothIndex(index_obj = ex_hpiindex,
                        order = 3,
                        in_place = TRUE)
}
