% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothIndex.R
\name{smoothSeries}
\alias{smoothSeries}
\title{smoothSeries}
\usage{
Lorem Ipsum...
}
\arguments{
\item{series}{Series to be smoothed}

\item{order}{Number of nearby period to smooth with}

\item{...}{Additional Arguments}
}
\value{
a list of objects ts and 'smooth_index` object with smoothed index
}
\description{
Smooths a series of indexex
}
\section{Further Details}{

Leaving order blank default to a moving average with order 3.
}

\examples{
a <- 1
}
