% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAccuracy.R
\name{calcAccuracy}
\alias{calcAccuracy}
\title{calcAccuracy}
\usage{
Lorem Ipsum...
}
\arguments{
\item{hpi_obj}{Object of class 'hpi'}

\item{series_name}{default = 'series'; name of the object in hpi_obj containing the series}

\item{test_method}{default = 'insample'; also 'kfold' or 'forecast'}

\item{test_type}{default = 'rs'; Type of data to use for test.  See details.}

\item{index_data}{= NULL; Extra data if the test_type doesn't match data in hpi_obj}

\item{...}{Additional Arguments}
}
\value{
hpimodel object
}
\description{
Estimate index Accuracy using one of a number of measures
}
\section{Further Details}{

'srs' test type tests the ability of the index to correctly predict the second value in a sale-resale pair (ABS)
'relhed' test type tests the ability of the index to improve an OLS model that doesn't account for time.  (REL)
}

