% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcForecastError.R
\name{calcForecastError}
\alias{calcForecastError}
\title{calcForecastError}
\usage{
Lorem Ipsum...
}
\arguments{
\item{is_obj}{Object of class 'hpiseries'}

\item{pred_df}{Set of sales to be used for predicitive quality of index}

\item{return_indexes}{Defaul = FALSE; return the forecasted indexes}

\item{...}{Additional Arguments}
}
\value{
x
}
\description{
Estimate out-of-sample index errors using a forecast method
}
\section{Further Details}{

If 'max_period' is left NULL, then it will forecast up to the end of the data
}

\examples{
a <- 1
}
