% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcForecastError.R
\name{calcForecastError}
\alias{calcForecastError}
\title{calcForecastError}
\usage{
hed_error <- calcForecastError(is_obj = hed_series, pred_df = rt_data)
}
\arguments{
\item{is_obj}{Object of class 'hpiseries'}

\item{pred_df}{Set of sales to be used for predicitive quality of index}

\item{return_forecasts}{default = FALSE; return the forecasted indexes}

\item{forecast_length}{default = 1; Length of period(s) in time to forecast}

\item{...}{Additional Arguments}
}
\value{
object of class `hpiaccuracy` inheriting from class `data.frame` containing the following fields:
\describe{
\item{prop_id}{Property Identification number}
\item{pred_price}{Predicted price}
\item{pred_error}{(Prediction - Actual) / Actual}
\item{pred_period}{Period of the prediction}
}
}
\description{
Estimate out-of-sample index errors using a forecast method
}
\section{Further Details}{

If you set `return_forecasts` = TRUE, the forecasted indexes for each period will be returned
in the `forecasts` attribute of the `hpiaccuracy` object. (attr(accr_obj, 'forecasts')

For now, the `pred_df` object must be a set of repeat transactions with the class `rt`,
inheriting from `hpidata`
}

\examples{
# Load example series
 data(ex_serieshpi)

# Load prediction data
 data(ex_rtdata)

# Calculate forecast accuracty
 fc_accr <- calcForecastError(is_obj = ex_serieshpi,
                              pred_df = ex_rtdata)
}
