% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIndexSeries.R
\name{calcIndexSeries}
\alias{calcIndexSeries}
\title{calcIndexSeries}
\usage{
Lorem Ipsum...
}
\arguments{
\item{hpi_obj}{Object of class 'hpi'}

\item{max_period}{Default=NULL; Maximum number of periods to forecast up to}

\item{name_prefix}{Default=NULL; Prefix to add before last time period if naming indexes}

\item{in_place}{Add to existing hpi object}

\item{in_place_name}{default = 'series'; Give it a different name than the default}

\item{...}{Additional Arguments}

\item{train_range}{Number of periods to use as purely training before forecast starts}
}
\value{
hpimodel object
}
\description{
Calculates a series of indexes, each one period longer than the previous
}
\section{Further Details}{

If 'max_period' is left NULL, then it will forecast up to the end of the data
}

\examples{
a <- 1
}
