% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIndexVolatility.R
\name{calcIndexVolatility}
\alias{calcIndexVolatility}
\title{calcIndexVolatility}
\usage{
calcIndexVolatility(index, window, in_place = FALSE, ...)
}
\arguments{
\item{index}{time-series index object}

\item{window}{periods over which to calculate the volatility}

\item{...}{Additional arguments}
}
\value{
A metric of volatility
}
\description{
Calculate index volatility
}
\section{Further Details}{

}

\examples{
index_vol <- calcIndexVolatility(index=index,
                                 window=3)
}
