% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcInSampleError.R
\name{calcInSampleError}
\alias{calcInSampleError}
\title{Calculate index errors in sample}
\usage{
calcInSampleError(pred_df, index, ...)
}
\arguments{
\item{pred_df}{Set of sales against which to test predictions}

\item{index}{Index (of class) to be tested for accuracy}

\item{...}{Additional Arguments}
}
\value{
object of class `hpiaccuracy` inheriting from class `data.frame`
containing the following fields:
\describe{
  \item{prop_id}{Property Identification number}
  \item{pred_price}{Predicted price}
  \item{pred_error}{(Prediction - Actual) / Actual}
  \item{pred_period}{Period of the prediction}
}
}
\description{
Estimate the predictive error of an index via an in-sample approach.
}
\section{Further Details}{

In addition to being a stand-alone function, it is also used by `calcForecastError`
and `calcKFoldError``
}

\examples{
# Load example data
  data(ex_rtdata)
  data(ex_hpiindex)

# Calculate accuracy
  in_accr <- calcInSampleError(pred_df = ex_rtdata,
                               index = ex_hpiindex$value)

}
