% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRevision.R
\name{calcRevision}
\alias{calcRevision}
\title{calcRevision}
\usage{
Lorem Ipsum...
}
\arguments{
\item{series_name}{default = 'series'; name of the series if extracting}

\item{in_place}{default = FALSE; calculating in place (adding to hpi)}

\item{in_place_name}{default = 'rev'; name of revision object in_place}

\item{...}{Additional Arguments}

\item{index_obj}{A list of progressively longer indexes (from calcForecstErrors())}
}
\value{
list containing:
\item{period} Mean revision by period
\item{total} Mean revision of all periods in all indexes
}
\description{
Estimate revision figures for a series of indexes
}
\section{Further Details}{

}

\examples{
a <- 1
}
