% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRevision.R
\name{calcRevision}
\alias{calcRevision}
\title{calcRevision}
\usage{
Lorem Ipsum...
}
\arguments{
\item{index_obj}{A list of progressively longer indexes (from calcForecstErrors())}

\item{...}{Additional Arguments}
}
\value{
list containing:
\item{period} Mean revision by period
\item{total} Mean revision of all periods in all indexes
}
\description{
Estimate revision figures for a series of indexes
}
\section{Further Details}{

}

\examples{
a <- 1
}
