% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateToPeriod.R
\name{checkDate}
\alias{checkDate}
\title{Validate the date argument}
\usage{
checkDate(x_date, name)
}
\arguments{
\item{x_date}{Date string or vector}

\item{name}{Name of argument to return in error/warning message}
}
\value{
Adjusted date field
}
\description{
Internal function to validate (or convert) the provided date field
}
\examples{

 # Load Data
  data(ex_sales)

 # Check date
  date_checked <- checkDate(x_date = ex_sales$sale_date,
                            name = 'sale date')
}
