% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSeries.R
\name{createSeries}
\alias{createSeries}
\title{createSeries}
\usage{
createSeries(hpi_obj, train_period = 12)
}
\arguments{
\item{hpi_obj}{Object of class 'hpi'}

\item{train_period}{Number of periods to use as purely training before creating indexes}

\item{max_period}{default=NULL; Maximum number of periods to create the index up to}

\item{...}{Additional Arguments}
}
\value{
An `hpiseries` object -- a list of `hpiindex` objects.
}
\description{
Calculates a series of indexes, each one period longer than the previous
}
\section{Further Details}{

`train_peried` Represents the shortest index that you will create. For certain approaches
, such as a repeat transaction model, indexes shorter than 10 will likely be highly unstable.

If 'max_period' is left NULL, then it will forecast up to the end of the data
}

\examples{
\dontrun{
hpi_series <- createSereis(hpi_obj = hpi_obj,
                           train_period = 12)
}
}
