% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateToPeriod.R
\name{dateToPeriod}
\alias{dateToPeriod}
\title{dateToPeriod}
\usage{
dateToPeriod(trans_df, date, periodicity = NULL, min_date = NULL,
  max_date = NULL, adj_type = "move", ...)
}
\arguments{
\item{trans_df}{data.frame of raw transactions}

\item{date}{name of field containing the date of the sale in Date or POSIXt format}

\item{periodicity}{type of periodicity to use ('yearly', 'quarterly', 'monthly' or 'weekly)}
}
\value{
data frame with three new fields:
trans_period: integer value counting from the minimum transaction date in the periodicity selected. Base value is 1. Primarily for modeling
date_value: float value of year and periodicty in numeric form (primarily for plotting)
date_name: text value of the period in the format, "Year-Period". (primarily for labeling)
}
\description{
Convert dates into time periods for use in sale-resale models
}
\section{Further Details}{

trans_period conat from the minimum transaction date provided.  As such the period counts
are relative, not absolute
Additionally, this function modifies the data.frame that it is given and return that same
data.frame that it is given and returns that data.frame with the new fields attached.
It does so because this function is not intended as a stand-alone function but rather
one to be called by the ***CreateSales set of functions with hpiR
}

