% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hedIndex.R
\name{hedIndex}
\alias{hedIndex}
\title{hedIndex}
\usage{
hedIndex(sales_df, date, price, sale_id, prop_id, estimator = "base",
  log_dep = TRUE, periodicity = "month", dep_var = NULL, ind_var = NULL,
  hed_spec = NULL, ...)
}
\arguments{
\item{sales_df}{data.frame of sale transactions}

\item{date}{Field contain the sales date}

\item{price}{Field contain the sale price}

\item{sale_id}{Field containing the unique sale identifier}

\item{prop_id}{Field containing the property identifier}

\item{estimator}{default = 'base', Type of estimator to use.  'base', 'robust' or 'weighted'}

\item{log_dep}{default = TRUE, Should the dependent variable (price) be logged?}

\item{periodicity}{default = 'month', Periodicity of time to estimate index at}

\item{dep_var}{default = NULL, dependent variable in hedonic model}

\item{ind_var}{default = NULL, independent variables in the hedonic model}

\item{hed_spec}{default = NULL, Full hedonic model specification}

\item{...}{Additional Arguments}
}
\value{
hpi object
}
\description{
Creates a house price index from a set of sales transactions using the
hedonic price index
}
\section{Further Details}{

}

\examples{
sea_hed_index <- rsIndex(sales_df = seattle_sales,
                        date = 'sale_date',
                        price = 'sale_price',
                        sale_id = 'uniq_id',
                        prop_id = 'pinx')
}
