% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hedIndex.R
\name{hedIndex}
\alias{hedIndex}
\title{hedIndex}
\usage{
hedIndex(trans_df, dep_var = NULL, ind_var = NULL, hed_spec = NULL, ...)
}
\arguments{
\item{trans_df}{data.frame of transactions}

\item{dep_var}{default = NULL, dependent variable in hedonic model}

\item{ind_var}{default = NULL, independent variables in the hedonic model}

\item{hed_spec}{default = NULL, Full hedonic model specification}

\item{...}{Additional Arguments}

\item{date}{Field contain the transaction date}

\item{price}{Field contain the transaction price}

\item{trans_id}{Field containing the unique transaction identifier}

\item{prop_id}{Field containing the property identifier}

\item{estimator}{default = 'base', Type of estimator to use.  'base', 'robust' or 'weighted'}

\item{log_dep}{default = TRUE, Should the dependent variable (price) be logged?}

\item{periodicity}{default = 'month', Periodicity of time to estimate index at}
}
\value{
hpi object
}
\description{
Creates a house price (rent) index from a set of transactions using the
hedonic price method
}
\section{Further Details}{

}

