% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hedModel.R
\name{hedModel.weighted}
\alias{hedModel.weighted}
\title{hedModel.weighted}
\usage{
\method{hedModel}{weighted}(estimator, hed_df, hed_spec, ...)
}
\section{Further Details}{

See `?hedModel` for more information
}

