% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hedModel.R
\name{hedModel.weighted}
\alias{hedModel.weighted}
\title{Hedonic model approach with weighted estimator}
\usage{
\method{hedModel}{weighted}(estimator, hed_df, hed_spec, ...)
}
\arguments{
\item{estimator}{Type of model to estimates (base, robust, weighted)}

\item{hed_df}{Repeat sales dataset from hedCreateSales()}

\item{hed_spec}{Model specification (`formula` object)}

\item{...}{Additional arguments}
}
\description{
Use of weighted estimator in hedonic model approach
}
\section{Further Details}{

See `?hedModel` for more information
}

