% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpiModel.R
\name{hpiModel.heddata}
\alias{hpiModel.heddata}
\title{hpiModel.heddata}
\usage{
Lorem Ipsum...
}
\arguments{
\item{hpi_df}{Dataset created by one of the *CreateSales() function in this package.}

\item{estimator}{Type of estimator to be used ('base', 'weighted', 'robust')}

\item{log_dep}{default=TRUE; should the dependent variable (change in price) be logged?}

\item{hed_spec}{default=NULL; hedonic model specification}

\item{dep_var}{default=NULL; dependent variable of the model}

\item{ind_var}{default=NULL; independent variable(s) of the model}

\item{trim_model}{default TRUE, should excess be trimmed from model results ('lm' or 'rlm' object)?}

\item{...}{Additional Arguments}
}
\value{
hpimodel object consisting of:
\item{estimator}
\item{coefficients}
\item{model_obj: class `rtmodel` or `hedmodel`}
\item{mod_spec}
\item{log_dep}
\item{base_price}
\item{periods: `data.frame` of periods}
\item{approach: type of model used}
}
\description{
Estimate the model for any method of house price index.  Generic method.
}
\examples{
# Load Data
 data(ex_heddata)

 # Create model object
 hpi_model <- hpiModel(hpi_df = ex_heddata,
                       estimator = 'base',
                       dep_var = 'price',
                       ind_var = c('tot_sf', 'beds', 'baths'),
                       log_dep = TRUE)
}
