% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpiModel.R
\name{hpiModel}
\alias{hpiModel}
\title{Wrapper to estimate model approaches (generic method)}
\usage{
hpiModel(model_type, hpi_df, estimator = "base", log_dep = TRUE,
  trim_model = TRUE, mod_spec = NULL, ...)
}
\arguments{
\item{model_type}{Type of model to estimate ('rt', 'hed', 'rf')}

\item{hpi_df}{Dataset created by one of the *CreateTrans() function in this package.}

\item{estimator}{Type of estimator to be used ('base', 'weighted', 'robust')}

\item{log_dep}{default TRUE, should the dependent variable (change in price) be logged?}

\item{trim_model}{default TRUE, should excess be trimmed from model results ('lm' or 'rlm' object)?}

\item{mod_spec}{Model specification}

\item{...}{Additional Arguments}
}
\value{
hpimodel object consisting of:
\describe{
\item{estimator}{Type of estimator}
\item{coefficients}{Data.frame of coefficient}
\item{model_obj}{class `rtmodel` or `hedmodel`}
\item{mod_spec}{Full model specification}
\item{log_dep}{Binary: is the dependent variable in logged format}
\item{base_price}{Mean price in the base period}
\item{periods}{`data.frame` of periods}
\item{approach}{Type of model used}
}
}
\description{
Generic method to estimate modeling approaches for indexes
}
\examples{

 # Load data
 data(ex_sales)

 # With a raw transaction data.frame
 rt_data <- rtCreateTrans(trans_df = ex_sales,
                          prop_id = 'pinx',
                          trans_id = 'sale_id',
                          price = 'sale_price',
                          periodicity = 'monthly',
                          date = 'sale_date')

 # Create model object
 hpi_model <- hpiModel(model_type = 'rt',
                       hpi_df = rt_data,
                       estimator = 'base',
                       log_dep = TRUE)

 # For custom weighted repeat transaction model

 hpi_model_wgt <- hpiModel(model_type = 'rt',
                           hpi_df = rt_data,
                           estimator = 'weighted',
                           weights = runif(nrow(rt_data), 0, 1))


}
