% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcKFoldError.R
\name{matchKFold}
\alias{matchKFold}
\title{matchKFoldData}
\usage{
Lorem Ipsum...
}
\arguments{
\item{train_df}{Data.frame of training data}

\item{pred_df}{Data to be used for prediction}
}
\value{
list
\item{train} Training data
\item{score} Scoring data
}
\description{
Makes specific selections of scoring data (Generic Method)
}
\section{Further Details}{

Lorem Ipsum...
}

\examples{
a <- 1
}
