% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plot.hpi}
\alias{plot.hpi}
\title{plot.hpi}
\usage{
Lorem Ipsum...
}
\arguments{
\item{hpi_obj}{Object of class HPI}

\item{show_imputed}{Highlight the imputed points}

\item{...}{Additional Arguments}
}
\value{
ggplot object
}
\description{
Simple Plot of an HPI object
}
\section{Further Details}{

Lorem Ipsum...
}

\examples{
a <- 1
}
