% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plot.hpi}
\alias{plot.hpi}
\title{Plot method for `hpi` object}
\usage{
\method{plot}{hpi}(x, ...)
}
\arguments{
\item{x}{Object to plot of class `hpi`}

\item{...}{Additional Arguments}
}
\value{
`plotindex` object inheriting from a ggplot object
}
\description{
Specific plotting method for hpi objects
}
\section{Further Details}{

Additional argument can include those argument for `plot.hpindex``
}

\examples{
# Load Data
data(ex_hpi)

# Plot data
plot(ex_hpi)
plot(ex_hpi, smooth = TRUE)
}
