% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plot.hpi}
\alias{plot.hpi}
\title{plot.hpi}
\usage{
plot(hpi_obj, show-impute)
}
\arguments{
\item{x}{Object to plot of class `hpi`}

\item{...}{Additional Arguments}
}
\value{
`plotindex` object inheriting from a ggplot object
}
\description{
Simple Plot of an HPI object
}
\section{Further Details}{

Additional argument can include those argument for `plot.hpindex``
}

\examples{
# Load Data
data(ex_hpi)

# Plot data
plot(ex_hpi)
plot(ex_hpi, smooth = TRUE)
}
