% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plot.hpiaccuracy}
\alias{plot.hpiaccuracy}
\title{Plot method for `hpiaccuracy` object}
\usage{
\method{plot}{hpiaccuracy}(x, return_plot = FALSE, do_plot = TRUE, ...)
}
\arguments{
\item{x}{Object to plot of class `hpiaccuracy``}

\item{return_plot}{default = FALSE; Return the plot to the function call}

\item{do_plot}{default = FALSE; Execute plotting to terminal/console}

\item{...}{Additional Arguments}
}
\value{
`plotaccuracy` object inheriting from a ggplot object
}
\description{
Specific plotting method for hpiaccuracy objects
}
\examples{
# Load data
 data(ex_hpiaccuracy)

# Make Plot
 plot(ex_hpiaccuracy)
}
