% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plot.hpiindex}
\alias{plot.hpiindex}
\title{plot.hpiindex}
\usage{
Lorem Ipsum...
}
\arguments{
\item{index_obj}{Object of class hpiindex}

\item{show_imputed}{Highlight the imputed points}

\item{...}{Additional Arguments}
}
\value{
ggplot object
}
\description{
Simple Plot of an hpiindex object
}
\section{Further Details}{

Lorem Ipsum...
}

\examples{
a <- 1
}
