% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plot.hpiindex}
\alias{plot.hpiindex}
\title{plot.hpiindex}
\usage{
plot(hpiindex_obj, show_imputed, smooth, ...)
}
\arguments{
\item{index_obj}{Object of class `hpiindex``}

\item{show_imputed}{defautl = FALSE; highlight the imputed points}

\item{smooth}{default = FALSE; plot the smoothed index}

\item{...}{Additional Arguments}
}
\value{
`plotindex` object inheriting from a ggplot object
}
\description{
Simple Plot of an hpiindex object
}
\examples{
# Load data
 data(ex_hpiindex)

# Make Plot
 plot(ex_hpiindex)
}
