% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plot.indexvolatility}
\alias{plot.indexvolatility}
\title{plot.indexvolatility}
\usage{
plot(x=vol_obj)
}
\arguments{
\item{x}{Object to plot of class `indexvolatility``}
}
\value{
`plotvolatility` object inheriting from a ggplot object
}
\description{
Simple Plot of an indexvolatility object
}
\examples{
# Load data
 data(ex_indexvolatility)

# Make Plot
 plot(ex_indexvolatility)
}
