% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plot.serieshpi}
\alias{plot.serieshpi}
\title{plot.serieshpi}
\usage{
plot(x)
}
\arguments{
\item{x}{Object of class `serieshpi``}
}
\value{
`plotseries` object inheriting from a ggplot object
}
\examples{
# Load data
 data(ex_serieshpi)

# Make Plot
 plot(ex_serieshpi)
}
