% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plot.seriesrevision}
\alias{plot.seriesrevision}
\title{Plot method for `seriesrevision` object}
\usage{
\method{plot}{seriesrevision}(x, measure = "median", ...)
}
\arguments{
\item{x}{Object to plot of class `seriesrevision`}

\item{measure}{default = 'median'; Metric to plot ('median' or 'mean')}

\item{...}{Additional Arguments}
}
\value{
`plotrevision` object inheriting from a ggplot object
}
\description{
Specific plotting method for seriesrevision objects
}
\examples{
# Load data
 data(ex_seriesrevision)

# Make Plot
 plot(ex_seriesrevision)
}
