% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfModel.R
\name{rfModel.base}
\alias{rfModel.base}
\title{Hedonic model approach with base estimator}
\usage{
\method{rfModel}{base}(estimator, rf_df, rf_spec, ...)
}
\arguments{
\item{estimator}{Type of model to estimates (base, robust, weighted)}

\item{...}{Additional arguments}
}
\description{
Use of base estimator in hedonic model approach
}
\section{Further Details}{

See `?rfModel` for more information
}

