% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsCreateSales.R
\name{rsCreateSales}
\alias{rsCreateSales}
\title{rsCreateSales}
\usage{
rsCreateSales(sales_df, prop_id, sale_id, price, date = NULL,
  periodicity = NULL, ...)
}
\arguments{
\item{sales_df}{sales transaction in either a data.frame or a
sales.df class from dateToPeriod() function}

\item{prop_id}{field contain the unique property identification}

\item{sale_id}{field containing the unique sale identification}

\item{price}{field containing the sale price}

\item{date}{default=NULL, field containing the date of the sale.
Only necessary if not passing a sales.df object}

\item{periodicity}{default=NULL, field containing the desired periodicity of analysis.
Only necessary if not passing a sales.df object}
}
\value{
data.frame of repeat sales. Note that a full data.frame of the possible
periods, their values and names can be found in the attributes to the returned `rs` object
}
\description{
Create repeat sale pairs from a data frame of sale transactions
}
\section{Further Details}{

Properties with greater than two sales during the period will make pairwise matches
among all sales.  Any property selling twice in the same period will remove the lower
priced of the two sales.
If passing a raw data.frame (not a 'sales.df' object) the "date" field should refer to
a field containing a vector of class POSIXt or Date.
}

\examples{
## With a raw data.frame
rep_sales <- rsCreateSales(sales_df = seattle_sales,
                           prop_id = 'pinx',
                           sale_id = 'uniq_id',
                           price = 'sale_price',
                           date = 'sale_date',
                           periodicity = 'month')

## When pre-calculating the time periods
sea_sales <- dateToPeriod(sales_df = seattle_sales,
                          date = 'sale_date',
                          periodicity = 'month')
rep_sales <- rsCreateSales(sales_df = sea_sales,
                           prop_id = 'pinx',
                           sale_id = 'uniq_id',
                           price = 'sale_price')

}
