% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtIndex.R
\name{rtIndex}
\alias{rtIndex}
\title{rtIndex}
\usage{
rtIndex(trans_df, ...)
}
\arguments{
\item{trans_df}{data.frame of transactions.  Can be a 'hpidata' or an 'rtdata' object.}

\item{...}{Additional Arguments}

\item{date}{Field contain the transaction date}

\item{price}{Field contain the transaction price}

\item{trans_id}{Field containing the unique transaction identifier}

\item{prop_id}{Field containing the property identifier}

\item{estimator}{default = 'base', Type of estimator to use.  'base', 'robust' or 'weighted'}

\item{log_dep}{default = TRUE, Should the dependent variable (price) be logged?}

\item{periodicity}{default = 'monthly', Periodicity of time to estimate index at}
}
\value{
hpi object
}
\description{
Creates a house price index from a set of ttransactions using the
repeat transactions method
}
\section{Further Details}{

}

\examples{
sea_rt_index <- rtIndex(trans_df = seattle_sales,
                        date = 'sale_date',
                        price = 'sale_price',
                        trans_id = 'uniq_id',
                        prop_id = 'pinx')
}
