% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtModel.R
\name{rtModel}
\alias{rtModel}
\title{rtModel}
\usage{
rtModel(rt_df, time_matrix, price_diff, estimator, ...)
}
\arguments{
\item{rt_df}{Repeat transactions dataset from rtCreateTrans()}

\item{time_matrix}{Time matrix object from rtTimeMatrix()}

\item{price_diff}{Difference in price betwen the two transactions}

\item{estimator}{Type of model to estimates (base, robust, weighted).  Must be in that class.}

\item{...}{Additional arguments}
}
\value{
`rtmod` object
}
\description{
Estimate repeat transactions model (method based on estimator class). Generic method.
}
\section{Further Details}{

}

\examples{
rt_model <- rtModel(rt_df = rt_sales,
                    time_matrix = time_matrix,
                    price_diff = price_diff,
                    estimator = structure('base', class='base'))
}
