% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtTimeMatrix.R
\name{rtTimeMatrix}
\alias{rtTimeMatrix}
\title{rtTimeMatrix}
\usage{
rtTimeMatrix(rtdata)
}
\arguments{
\item{rt_df}{object of class `rtdata`: repeat transaction data.frame created by rtCreateTrans()}
}
\value{
matrix to be used on the right hand side of a repeat sales regression model
}
\description{
Create a time matrix for a repeat transaction regression model
}
\section{Further Details}{

Time periods are calculated from the data provided.
}

\examples{
# Load Data
data(ex_rtdata)

# Create Matrix
rt_matrix <- rtTimeMatrix(ex_rtdata)
}
