% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothIndex.R
\name{smoothIndex}
\alias{smoothIndex}
\title{smoothIndex}
\usage{
Lorem Ipsum...
}
\arguments{
\item{index}{Index to be smoothed}

\item{order}{Number of nearby period to smooth with, multiple means multiple iterations}

\item{in_place}{default = FALSE; adds volatility metric to the `hpiindex` object}

\item{in_place_name}{default = 'vol'; name of volatility object in `hpiindex` object}

\item{...}{Additional Arguments}
}
\value{
a ts and 'smooth_index` object with smoothed index
}
\description{
Smooths an index
}
\section{Further Details}{

Leaving order blank default to a moving average with order 3.
}

\examples{
a <- 1
}
