% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothIndex.R
\name{smoothSeries}
\alias{smoothSeries}
\title{Smooth all indexes in a series}
\usage{
smoothSeries(series_obj, order = 3, ...)
}
\arguments{
\item{series_obj}{Series to be smoothed}

\item{order}{Number of nearby period to smooth with}

\item{...}{Additional Arguments}
}
\value{
a `serieshpi` object with a smoothed index in each `hpiindex` object
}
\description{
Smoothes all indexes within a progressive series of indexes
}
\section{Further Details}{

Leaving order blank default to a moving average with order 3.
}

\examples{
# Load Data
data(ex_serieshpi)

# Smooth indexes
sm_series <- smoothSeries(series_obj = ex_serieshpi,
                          order = 5)
}
